Widen multi_tensor_apply tensor sizes to int64
TensorListMetadataBase::sizes was declared int32 but populated from
Tensor::numel(), so a tensor with numel > INT_MAX truncated to a negative
size and the consumer kernels then computed out-of-bounds offsets, hitting
an illegal memory access at the next sync.

Store sizes as int64_t and read them into an int64_t (or the existing
index_t on the already-templated Adam kernels) before the
n -= chunk_idx * chunk_size subtraction. Widen the chunk_size kernel
argument in the non-Adam consumers to int64_t as well so the
chunk_idx * chunk_size element offset is computed in 64-bit.

Fixes #2918

+++ transformer_engine/common/multi_tensor/adam.cu
@@ -382,7 +382,7 @@ struct AdamFunctor {
 
 template <typename T, typename FULL_T, typename MOMENT_T>
 struct AdamCapturableFunctor {
-  __device__ __forceinline__ void operator()(int chunk_size, volatile int *noop_gmem,
+  __device__ __forceinline__ void operator()(int64_t chunk_size, volatile int *noop_gmem,
                                              TensorListMetadata<4> &tl,  // NOLINT(*)
                                              const float beta1, const float beta2, const int *step,
                                              const int bias_correction, const float epsilon,
@@ -402,7 +402,7 @@ struct AdamCapturableFunctor {
     // int tensor_num = tl.start_tensor_this_launch + tensor_loc;
 
     int chunk_idx = tl.block_to_chunk[blockIdx.x];
-    int n = tl.sizes[tensor_loc];
+    int64_t n = tl.sizes[tensor_loc];
 
     T *g = reinterpret_cast<T *>(tl.addresses[0][tensor_loc]);
     g += chunk_idx * chunk_size;
@@ -476,7 +476,7 @@ struct AdamCapturableFunctor {
 
 template <typename T, typename FULL_T, typename MOMENT_T>
 struct AdamCapturableMasterFunctor {
-  __device__ __forceinline__ void operator()(int chunk_size, volatile int *noop_gmem,
+  __device__ __forceinline__ void operator()(int64_t chunk_size, volatile int *noop_gmem,
                                              TensorListMetadata<5> &tl,  // NOLINT(*)
                                              const float beta1, const float beta2, const int *step,
                                              const int bias_correction, const float epsilon,
@@ -496,7 +496,7 @@ struct AdamCapturableMasterFunctor {
     // int tensor_num = tl.start_tensor_this_launch + tensor_loc;
 
     int chunk_idx = tl.block_to_chunk[blockIdx.x];
-    int n = tl.sizes[tensor_loc];
+    int64_t n = tl.sizes[tensor_loc];
 
     T *g = reinterpret_cast<T *>(tl.addresses[0][tensor_loc]);
     g += chunk_idx * chunk_size;
